flight_api/requirements.txt
attrs==19.3.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
idna==2.10
importlib-metadata==1.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
more-itertools==8.4.0
packaging==20.4
pluggy==0.13.1
psycopg2-binary==2.8.5
py==1.9.0
pyparsing==2.4.7
pytest==5.4.3
requests==2.24.0
six==1.15.0
urllib3==1.25.9
wcwidth==0.2.5
Werkzeug==1.0.1
zipp==3.1.0
